decimal now a direct wrapper of double for less memory consumption

+++ invesdwin-util/src/main/java/de/invesdwin/util/math/decimal/scaled/IDecimalScale.java
@@ -1,11 +1,10 @@
 package de.invesdwin.util.math.decimal.scaled;
 
 import de.invesdwin.util.math.decimal.AScaledDecimal;
-import de.invesdwin.util.math.decimal.Decimal;
 
 public interface IDecimalScale<T extends AScaledDecimal<T, S>, S extends IDecimalScale<T, S>> {
 
-    Decimal convertValue(T parent, Decimal value, S scale);
+    double convertValue(T parent, double value, S scale);
 
     int getDefaultDecimalDigits(T parent);
